Refactor PageParameterSchema to new domain module
We want to move the concerns around page parameters to a new domain
module to reduce the size and the Page domain module. This will make
adding functionality around page parameters easier to manage while
reducing complexity for the core page domain module.
